Refactoring ValueLength-Component
