+++ src/components/Home/Home.test.tsx
@@ -13,17 +13,6 @@ describe('Home component tests', () => {
     expect(passwordLengthSpan).toBeInTheDocument();
   });
 
-  test('It updates word length counter', () => {
-    const password = 'test-password';
-
-    const { getByTestId } = render(<Home/>);
-    const passwordInput = getByTestId('password-input');
-    fireEvent.change(passwordInput,  { target: { value: password } });
-    const passwordLengthSpan = getByTestId('password-length');
-
-    expect(passwordLengthSpan).toHaveTextContent(`${password.length}`);
-  });
-
   test('It matches all rules', () => {
     const password = 'test-password';
