+++ src/components/ValueLength/ValueLength.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface ValueLengthProps {
+  value: string
+}
+
+const ValueLength: React.FunctionComponent<ValueLengthProps> = ({ value }) => {
+  return (
+    <span data-testid="password-length">{value.length}</span>
+  );
+};
+
+export default ValueLength;
