+++ src/components/ValueLength/ValueLengthTest.tsx
@@ -1,0 +1,17 @@
+import { render, fireEvent } from '@testing-library/react';
+import ValueLength from './ValueLength';
+
+describe('ValueLength component tests', () => {
+
+  test('It updates word length counter', () => {
+    const password = 'test-password';
+
+    const { getByTestId } = render(<ValueLength value={password}/>);
+    const passwordInput = getByTestId('password-input');
+    fireEvent.change(passwordInput,  { target: { value: password } });
+    const passwordLengthSpan = getByTestId('password-length');
+
+    expect(passwordLengthSpan).toHaveTextContent(`${password.length}`);
+  });
+
+});
